updates to work with mantle
