+++ hooks/useImportSafes.tsx
@@ -1,14 +1,16 @@
+import type { UserGnosisSafe } from '@charmverse/core/prisma';
 import type { Signer } from 'ethers';
 import log from 'loglevel';
 import { useCallback, useState } from 'react';
 
-import { importSafesFromWallet } from 'lib/gnosis/gnosis.importSafes';
+import charmClient from 'charmClient';
+import { getSafesForAddresses } from 'lib/gnosis/gnosis';
 
 import useMultiWalletSigs from './useMultiWalletSigs';
 import { useSnackbar } from './useSnackbar';
 import { useUser } from './useUser';
 
-export default function useImportSafes() {
+export function useImportSafes() {
   const { data, mutate } = useMultiWalletSigs();
   const { user } = useUser();
   const { showMessage } = useSnackbar();
@@ -41,4 +43,27 @@ export default function useImportSafes() {
   };
 
   return { isLoadingSafes, importSafes };
+}
+
+interface ImportSafeProps {
+  signer: Signer;
+  addresses: string[];
+  getWalletDetails?: (address: string) => UserGnosisSafe | null | undefined;
+}
+
+export async function importSafesFromWallet({ signer, addresses, getWalletDetails }: ImportSafeProps) {
+  const safes = await getSafesForAddresses(signer, addresses);
+
+  const safesData = safes.map((safe) => ({
+    address: safe.address,
+    owners: safe.owners,
+    threshold: safe.threshold,
+    chainId: safe.chainId,
+    name: getWalletDetails?.(safe.address)?.name ?? null, // get existing name if user gave us one
+    isHidden: getWalletDetails?.(safe.address)?.isHidden ?? false
+  }));
+
+  await charmClient.gnosisSafe.setMyGnosisSafes(safesData);
+
+  return safes.length;
 }
